requirements.txt
Flask==2.3.3
pdfplumber==0.10.3
google-generativeai==0.8.3